vix: the fixture store leaves core, and tree names come from the seam
The move, second half: FixtureStore — the offline filesystem store, its
virtual-file rerun overlay, and its origin declaration — relocates from
vix-core/src/runtime/fixture.rs to vixen-runtime/src/fixture.rs, joining
the surface declarations so the harness crate owns the store, the overlay,
the spellings, and the fixture-tree\0 namespace in ONE module. The fixture
root moves with it (vix-core/tests/fixtures -> vixen-runtime/tests/fixtures)
because the root is DECLARED now: the store's default resolves against the
harness crate's own manifest dir, and no CARGO_MANIFEST_DIR default is
resident in any Runtime. The manifest lock test rides along.

fixture_tree_name dies with the sentinel it spelled. Its role — building
name-relative projection paths (<name>/<path>) for lazily-backed trees —
was the last backend knowledge inside the tree-read/tree-glob primitives,
so it becomes seam vocabulary: OriginAdapterSet::tree_handle_name derives a
handle's adapter-relative name from the DECLARED namespaces (strip the
owning declaration's prefix), surfaced to primitives as
EffectCtx::tree_handle_name and CodataDrainCtx::source_origin_name. Paths
stay name-relative rather than going handle-relative, deliberately:
projection strings are identity-bearing (witness paths, never_read
contracts like p"small-crate/src/lib.rs", directory-observation digests),
and the name IS the handle's coordinate in its adapter's space — the seam
deriving it from declarations is what the note's namespace amendment is
for. Byte-identical paths, witnesses, and rung behavior; the whole corpus
stays green.

vix-core keeps only the seam (origin.rs, TreeEntryKind, the adapter
traits). Core tests stop depending on the moved store, argued per case:
- tree_resident's ustar/carrier tests exercise CORE representation code, so
  they keep core-owned bytes: the 4KB sample tar checks in as
  src/runtime/testdata (the harness still ships the same archive in its
  registry, hash-pinned there).
- scheduler's witness-reverification tests exercise the SEAM, so they get
  an in-memory OverlayTreeOrigin with a declared test-tree\0 namespace and
  a mutable-world switch — a sharper test than driving a filesystem store,
  and core spells no fixture namespace.

Remaining fixture spellings in core are now exactly: the retire-and-reserve
history markers (ordinals 90/94, WireArg slot 2, FixtureEntryKind's
retirement note), the binding tombstone test pinning that the names do NOT
resolve, and the transitional REGISTRY_MANIFEST_COORDINATE — which the next
commit retires with ReadProjection::RegistryManifest.
