+++ vix-core/src/runtime/mod.rs
@@ -11,7 +11,6 @@ mod blob_persistence;
 mod decode_value;
 mod error;
 mod exec_backend;
-mod fixture;
 mod identity;
 mod model;
 mod observe;
@@ -28,7 +27,6 @@ pub use blob_persistence::*;
 pub use decode_value::*;
 pub use error::*;
 pub use exec_backend::*;
-pub use fixture::*;
 pub use identity::*;
 pub use model::*;
 pub use observe::*;
