+++ vixen-runtime/tests/fixtures/elf/libtiny-x86_64.so
