+++ vixen-runtime/tests/fixtures/sparse-index/snapshot-2025-03-04/fi/nd/find-msvc-tools
