+++ vixen-runtime/tests/fixtures/sparse-index/snapshot-2025-03-04/it/oa/itoa
